Add cloud api authenticatication mechanism
Up to now the VcdAPiClient was using only `/api/session` endpoint for
the authentication and the corresponding API model.
This endpoint is deprecated and the openapi should be used instead.
This CLN introduces that.

Note: `@vcd/bindings` are not updated yet, so we have manually added the
`openapi` bindings

Testing done using a plugin:
0. Without doing any explicit backend call verify that the user can
access `cloudApiXXX` properties as well as their api/session counterparts.

1. Perform a `vcdApi.cloudApiLogin()` with a different user and verify
that the session, username, organization, can be retrieved through the
`cloudApiXXX` properties
2. perform a failing `vcdApi.cloudApiLogin()` attempt and verify that the session is nullified
3. get `cloudApiLocation` and verify they are retrieved from the backend just once

Signed-off-by: Ivo Rahov <irahov@vmware.com>