Add ability VcdApiClient to work with absolute URL and specify headers ofr CRUD requests
Till now the VcdApiClient was always prepending the baseURL as defined in the host
application (VCD). But there are cases when the URL is already an absolute one.
This CLN allows VcdApiClient to work with absolute URL.

While there, all the CRUD operations are enabled to get custom headers

Testing done: Use customize-portal plugin that was working with absolute URL and therefore
needed to use angular http client directly. With this CLN the plugin is simplified
and uses the VcdApiClient directly

Signed-off-by: Ivo Rahov <irahov@vmware.com>